requirements.txt
cfgv==3.3.1
distlib==0.3.5
Django==2.2
filelock==3.7.1
identify==2.5.2
m3-builder==1.2.0
m3-core==2.2.24
m3-django-compat==1.9.2
m3-objectpack==2.2.47
m3-ui==2.2.105
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.15.1
